_function/msteams-alert-function/requirement.txt
functions-framework==3.*
requests==2.*
google-cloud-logging==3.*